[FIX] l10n_fr_reports/saas~18.5: expected ids in migration script
replace_record_references_batch and execute expect ids but it was
possible for them to not having any which would break the upgrade
script.

closes odoo/upgrade#8379

Related: odoo/enterprise#93886

+++ migrations/l10n_fr_reports/saas~18.5.1.2/post-migrate.py
@@ -51,8 +51,9 @@ def migrate(cr, version):
         """
     cr.execute(query)
     ids = dict(cr.fetchall())
-    util.replace_record_references_batch(cr, ids, "account.report.async.export")
-    cr.execute("DELETE FROM account_report_async_export WHERE id IN %s", [tuple(ids.keys())])
+    if ids:
+        util.replace_record_references_batch(cr, ids, "account.report.async.export")
+        cr.execute("DELETE FROM account_report_async_export WHERE id IN %s", [tuple(ids.keys())])
 
     util.remove_column(cr, "account_report_async_document", "_upg_export_id")
